Vision code now canonical (probably)
I've got two repositories here. One is local. It reads from the file system and processes images so I can develop quickly and with good logging. The other is this one. Everything (mostly) works in the private repo but not here. I think I've fixed that by just copying the code over.