Candy shooter code
I wrote code for our candy shooter robot in Kotlin. Previously, Brandon had written code in blocks, but I can't find it and would like to have code in Kotlin so we can save it with the rest of our code. I hope to test this when it is time to use the candy shooter.